Delete smartphone only by admin user

+++ routers/router_smartphones.py
@@ -4,6 +4,10 @@ from sqlalchemy import func
 from sqlalchemy.orm import Session
 from classes.database import get_cursor
 from classes import models_orm, schemas_dto
+import utilities
+from typing import Annotated
+from fastapi.security import OAuth2PasswordBearer
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="/auth")
 
 router = APIRouter(
     prefix='/api/smartphones',
@@ -18,7 +22,7 @@ async def get_smartphones(
     all_phones = cursor.query(models_orm.Smartphones).limit(limit).offset(offset).all() # Lancement de la requête
     products_count= cursor.query(func.count(models_orm.Smartphones.id)).scalar()
     return {
-        "products": all_phones,
+        "smartphones": all_phones,
         "limit": limit,
         "total": products_count,
         "skip":offset
@@ -55,20 +59,30 @@ async def create_phone(payload: schemas_dto.Smartphone_POST_Body, cursor:Session
     cursor.refresh(new_phone)
     return {"message" : f"New Smartphone {new_phone.brand + ' ' + new_phone.model} added sucessfully with id: {new_phone.id}"} 
 
-# DELETE 
+
+# DELETE (Only possible by admin)
 @router.delete('/{phone_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_phone(phone_id:int, cursor:Session=Depends(get_cursor)):
+async def delete_phone(phone_id: int, token: str = Depends(oauth2_scheme), cursor: Session = Depends(get_cursor)):
+    decoded_user_id = utilities.decode_token(token)
+    connected_user = cursor.query(models_orm.Users).filter(models_orm.Users.id == decoded_user_id).first()
     # Recherche si le produit existe
     corresponding_product = cursor.query(models_orm.Smartphones).filter(models_orm.Smartphones.id == phone_id)
-    if(corresponding_product.first()):
+    # Check if user is admin
+    print(f'IS_ADMIN {connected_user.is_admin}')
+    if not connected_user.is_admin:
+        raise HTTPException(
+            status_code=status.HTTP_403_FORBIDDEN,
+            detail='Method allowed for admin only'
+        )
+    if corresponding_product.first() and connected_user.is_admin:
         # Continue to delete
-        corresponding_product.delete() # supprime
-        cursor.commit() # commit the stated changes (changement latent)
+        corresponding_product.delete()  # supprime
+        cursor.commit()  # commit the stated changes (changement latent)
         return
-    else: 
-        raise HTTPException (
+    else:
+        raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail=f'Ne corresponding product with id: {phone_id}'
+            detail=f'No corresponding product with id: {phone_id}'
         )
 
 # Update
